Cleanup test tagging
see #313.
