+++ libflatsurf/test/chain.test.cc
@@ -2,7 +2,7 @@
  *  This file is part of flatsurf.
  *
  *        Copyright (C) 2019-2020 Vincent Delecroix
- *        Copyright (C) 2019-2020 Julian Rüth
+ *        Copyright (C) 2019-2022 Julian Rüth
  *
  *  Flatsurf is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -21,11 +21,11 @@
 #include "../flatsurf/chain.hpp"
 #include "../flatsurf/flat_triangulation.hpp"
 #include "../flatsurf/half_edge.hpp"
-#include "external/catch2/single_include/catch2/catch.hpp"
+#include "cereal.helpers.hpp"
 #include "surfaces.hpp"
 
 namespace flatsurf::test {
-TEMPLATE_TEST_CASE("Chain Arithmetic", "[chain][arithmetic]", (long long), (mpz_class), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::NumberField>)) {
+TEMPLATE_TEST_CASE("Chain Arithmetic", "[Chain][operator+=][operator+][operator-][operator==][operator!=]", (long long), (mpz_class), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::NumberField>)) {
   using R2 = Vector<TestType>;
   auto square = makeSquare<R2>();
 
@@ -55,6 +55,13 @@ TEMPLATE_TEST_CASE("Chain Arithmetic", "[chain][arithmetic]", (long long), (mpz_
     REQUIRE(a * 1337 == a * 1336 + a);
     REQUIRE((a + b) * 1337 == a * 1337 + b * 1337);
   }
+}
+
+TEMPLATE_TEST_CASE("Serialization of a Chain", "[Chain][save][load]", (long long), (mpz_class), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::RationalField>), (exactreal::Element<exactreal::NumberField>)) {
+  using R2 = Vector<TestType>;
+  auto square = makeSquare<R2>();
+
+  testRoundtrip(Chain(*square, HalfEdge(1)));
 }
 
 }  // namespace flatsurf::test
