+++ libflatsurf/test/flat_triangulation_collapsed.test.cc
@@ -36,7 +36,7 @@
 
 namespace flatsurf::test {
 
-TEMPLATE_TEST_CASE("Collapse in a Flat Triangulation", "[flat_triangulation_collapsed][collapse]", (long long), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::NumberField>)) {
+TEMPLATE_TEST_CASE("Collapse in a Flat Triangulation", "[FlatTriangulationCollapsed]", (long long), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::NumberField>)) {
   using T = TestType;
   using R2 = Vector<TestType>;
