+++ libflatsurf/test/flow_decomposition.test.cc
@@ -44,7 +44,7 @@
 
 namespace flatsurf::test {
 
-TEST_CASE("Flow Decompositions Over Number Fields", "[flow_decomposition]") {
+TEST_CASE("Flow Decompositions Over Number Fields", "[FlowDecomposition]") {
   using T = renf_elem_class;
 
   SECTION("A Complicated Direction For A Veech Cathedral") {
@@ -64,7 +64,7 @@ TEST_CASE("Flow Decompositions Over Number Fields", "[flow_decomposition]") {
   }
 }
 
-TEMPLATE_TEST_CASE("Flow Decomposition", "[flow_decomposition]", (long long), (mpz_class), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::RationalField>), (exactreal::Element<exactreal::NumberField>)) {
+TEMPLATE_TEST_CASE("Flow Decomposition", "[FlowDecomposition]", (long long), (mpz_class), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::RationalField>), (exactreal::Element<exactreal::NumberField>)) {
   using T = TestType;
 
   const auto surface = GENERATE_SURFACES(T);
@@ -197,6 +197,145 @@ TEMPLATE_TEST_CASE("Flow Decomposition", "[flow_decomposition]", (long long), (m
       for (auto& component : flowDecomposition.components())
         REQUIRE(component.decomposition() == flowDecomposition);
     }
+  }
+}
+
+TEST_CASE("Parabolic", "[FlowDecomposition][parabolic]") {
+  SECTION("L with sqrt(3) horizontal") {
+    using T = renf_elem_class;
+    using R2 = Vector<T>;
+    auto surface = makeLParabolicNonParabolic<R2>();
+    CAPTURE(*surface);
+
+    const auto direction = Vector<T>(1, 0);
+    auto flowDecomposition = FlowDecomposition<FlatTriangulation<T>>(surface->clone(), direction);
+    CAPTURE(flowDecomposition);
+
+    REQUIRE(flowDecomposition.decompose());
+
+    REQUIRE(flowDecomposition.completelyPeriodic() == boost::logic::tribool(true));
+    REQUIRE(flowDecomposition.parabolic() == boost::logic::tribool(true));
+  }
+
+  SECTION("L with sqrt(3) vertical") {
+    using T = renf_elem_class;
+    using R2 = Vector<T>;
+    auto surface = makeLParabolicNonParabolic<R2>();
+    CAPTURE(*surface);
+
+    const auto direction = Vector<T>(0, 1);
+    auto flowDecomposition = FlowDecomposition<FlatTriangulation<T>>(surface->clone(), direction);
+    CAPTURE(flowDecomposition);
+
+    REQUIRE(flowDecomposition.decompose());
+
+    REQUIRE(flowDecomposition.completelyPeriodic() == boost::logic::tribool(true));
+    REQUIRE(flowDecomposition.parabolic() == boost::logic::tribool(false));
+  }
+
+  SECTION("L with sqrt(3) diagonal") {
+    using T = renf_elem_class;
+    using R2 = Vector<T>;
+    auto surface = makeLParabolicNonParabolic<R2>();
+    CAPTURE(*surface);
+
+    const auto direction = Vector<T>(1, 1);
+    auto flowDecomposition = FlowDecomposition<FlatTriangulation<T>>(surface->clone(), direction);
+    CAPTURE(flowDecomposition);
+
+    REQUIRE(flowDecomposition.decompose());
+
+    REQUIRE(flowDecomposition.completelyPeriodic() == boost::logic::tribool(true));
+    REQUIRE(flowDecomposition.parabolic() == boost::logic::tribool(false));
+  }
+
+  SECTION("(2,3,5) triangle in (1, 0)") {
+    using T = renf_elem_class;
+    using R2 = Vector<T>;
+    auto surface = make235<R2>();
+    CAPTURE(*surface);
+    CAPTURE(surface->area());
+
+    const auto direction = Vector<T>(1, 0);
+    auto flowDecomposition = FlowDecomposition<FlatTriangulation<T>>(surface->clone(), direction);
+    CAPTURE(flowDecomposition);
+
+    REQUIRE(flowDecomposition.decompose());
+
+    REQUIRE(flowDecomposition.completelyPeriodic() == boost::logic::tribool(true));
+    REQUIRE(flowDecomposition.components().size() == 4);
+    REQUIRE(flowDecomposition.parabolic() == boost::logic::tribool(true));
+  }
+
+  SECTION("(2,3,5) triangle in (0, 1)") {
+    using T = renf_elem_class;
+    using R2 = Vector<T>;
+    auto surface = make235<R2>();
+    CAPTURE(*surface);
+
+    const auto direction = Vector<T>(0, 1);
+    auto flowDecomposition = FlowDecomposition<FlatTriangulation<T>>(surface->clone(), direction);
+    CAPTURE(flowDecomposition);
+
+    REQUIRE(flowDecomposition.decompose());
+
+    REQUIRE(flowDecomposition.completelyPeriodic() == boost::logic::tribool(true));
+    REQUIRE(flowDecomposition.components().size() == 6);
+    REQUIRE(flowDecomposition.parabolic() == boost::logic::tribool(false));
+  }
+
+  SECTION("(2,3,5) triangle in (2c^2 - 1, -c^3 + 4c)") {
+    using T = renf_elem_class;
+    using R2 = Vector<T>;
+    auto surface = make235<R2>();
+    CAPTURE(*surface);
+
+    auto c = Q->gen();
+    const auto direction = Vector(2 * c * c - 1, -c * c * c + 4 * c);
+    auto flowDecomposition = FlowDecomposition<FlatTriangulation<T>>(surface->clone(), direction);
+    CAPTURE(flowDecomposition);
+
+    REQUIRE(flowDecomposition.decompose());
+
+    REQUIRE(flowDecomposition.completelyPeriodic() == boost::logic::tribool(true));
+    REQUIRE(flowDecomposition.components().size() == 4);
+    REQUIRE(flowDecomposition.parabolic() == boost::logic::tribool(true));
+  }
+
+  SECTION("(2,3,5) triangle in (c^2 - 5, c)") {
+    using T = renf_elem_class;
+    using R2 = Vector<T>;
+    auto surface = make235<R2>();
+    CAPTURE(*surface);
+
+    auto c = Q->gen();
+    const auto direction = Vector<T>(c * c - 5, c);
+    auto flowDecomposition = FlowDecomposition<FlatTriangulation<T>>(surface->clone(), direction);
+    CAPTURE(flowDecomposition);
+
+    REQUIRE(flowDecomposition.decompose());
+
+    REQUIRE(flowDecomposition.completelyPeriodic() == boost::logic::tribool(true));
+    REQUIRE(flowDecomposition.components().size() == 6);
+    REQUIRE(flowDecomposition.parabolic() == boost::logic::tribool(false));
+  }
+
+  SECTION("(2,3,5) triangle in (-5, -c^3 + 2c)") {
+    using T = renf_elem_class;
+    using R2 = Vector<T>;
+    auto surface = make235<R2>();
+    CAPTURE(*surface);
+
+    auto c = Q->gen();
+    const auto direction = Vector<T>(-5, -c * c * c + 2 * c);
+    auto flowDecomposition = FlowDecomposition<FlatTriangulation<T>>(surface->clone(), direction);
+    CAPTURE(flowDecomposition);
+
+    REQUIRE(flowDecomposition.decompose());
+
+    REQUIRE(flowDecomposition.completelyPeriodic() == boost::logic::tribool(true));
+    REQUIRE(flowDecomposition.components().size() == 6);
+    REQUIRE(flowDecomposition.parabolic() == boost::logic::tribool(false));
   }
 }
