+++ libflatsurf/test/path.test.cc
@@ -34,7 +34,7 @@ namespace flatsurf::test {
 
 using namespace flatsurf;
 
-TEMPLATE_TEST_CASE("Nullhomotopic Paths", "[path]", (long long), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::NumberField>)) {
+TEMPLATE_TEST_CASE("Nullhomotopic Paths", "[Path]", (long long), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::NumberField>)) {
   SECTION("Trivial Paths can be Tigthened") {
     using T = TestType;
     using Surface = FlatTriangulation<T>;
@@ -78,7 +78,7 @@ TEMPLATE_TEST_CASE("Nullhomotopic Paths", "[path]", (long long), (mpq_class), (r
   }
 }
 
-TEMPLATE_TEST_CASE("Short Paths", "[path]", (long long), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::NumberField>)) {
+TEMPLATE_TEST_CASE("Short Paths", "[Path]", (long long), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::NumberField>)) {
   SECTION("Tightening a Path Along two Edges of a Face") {
     using T = TestType;
     using Surface = FlatTriangulation<T>;
@@ -99,7 +99,7 @@ TEMPLATE_TEST_CASE("Short Paths", "[path]", (long long), (mpq_class), (renf_elem
   }
 }
 
-TEMPLATE_TEST_CASE("Random Paths", "[path]", (long long), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::NumberField>)) {
+TEMPLATE_TEST_CASE("Random Paths", "[Path]", (long long), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::NumberField>)) {
   SECTION("Tightening a Random Path") {
     using T = TestType;
     using Surface = FlatTriangulation<T>;
