+++ libflatsurf/test/vertex.test.cc
@@ -1,13 +1,13 @@
 /**********************************************************************
  *  This file is part of flatsurf.
  *
- *        Copyright (C) 2020 Julian Rüth
+ *        Copyright (C) 2020-2022 Julian Rüth
  *
  *  Flatsurf is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
  *  the Free Software Foundation, either version 3 of the License, or
  *  (at your option) any later version.
- *
+*
  *  Flatsurf is distributed in the hope that it will be useful,
  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -19,12 +19,12 @@
 
 #include "../flatsurf/edge.hpp"
 #include "../flatsurf/vertex.hpp"
-#include "external/catch2/single_include/catch2/catch.hpp"
+#include "cereal.helpers.hpp"
 #include "surfaces.hpp"
 
 namespace flatsurf::test {
 
-TEST_CASE("Vertices And Modifications of Surfaces", "[vertex][flip]") {
+TEST_CASE("Vertices And Modifications of Surfaces", "[Vertex][flip][source][target]") {
   auto hexagon = makeHexagonCombinatorial();
 
   REQUIRE(hexagon->vertices().size() == 2);
@@ -41,6 +41,11 @@ TEST_CASE("Vertices And Modifications of Surfaces", "[vertex][flip]") {
       REQUIRE(hexagon->vertices().size() == 2);
     }
   }
+}
+
+TEST_CASE("Serialization of a Vertex", "[Vertex][save][load]") {
+  auto square = makeSquare<Vector<long long>>();
+  testRoundtrip(square->vertices()[0]);
 }
 
 }  // namespace flatsurf::test
