+++ libflatsurf/test/vertical.test.cc
@@ -1,8 +1,7 @@
 /**********************************************************************
  *  This file is part of flatsurf.
  *
- *        Copyright (C)      2020 Vincent Delecroix
- *        Copyright (C) 2020-2022 Julian Rüth
+ *        Copyright (C) 2022 Julian Rüth
  *
  *  Flatsurf is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -18,27 +17,17 @@
  *  along with flatsurf. If not, see <https://www.gnu.org/licenses/>.
  *********************************************************************/
 
-#include "../flatsurf/edge.hpp"
-#include "../flatsurf/half_edge.hpp"
+#include "../flatsurf/vertical.hpp"
+
 #include "cereal.helpers.hpp"
 
 namespace flatsurf::test {
 
-TEST_CASE("Edge Basics", "[Edge][constructor][operator==][fromIndxe]") {
-  const int id = GENERATE(range(1, 1024));
+TEMPLATE_TEST_CASE("Serialization of a Vertical", "[Vertical][load][save]", (long long), (mpz_class), (mpq_class), (renf_elem_class), (exactreal::Element<exactreal::IntegerRing>), (exactreal::Element<exactreal::RationalField>), (exactreal::Element<exactreal::NumberField>)) {
+  using R2 = Vector<TestType>;
+  auto square = makeSquare<R2>();
 
-  const Edge e(id);
-
-  REQUIRE(e == Edge(id));
-  REQUIRE(-e.negative() == e.positive());
-  REQUIRE(e == HalfEdge(-id));
-  REQUIRE(e == HalfEdge(id));
-  REQUIRE(e.index() == id - 1);
-  REQUIRE(Edge::fromIndex(e.index()) == e);
+  testRoundtrip(Vertical(*square, square->fromHalfEdge(HalfEdge(1))));
 }
 
-TEST_CASE("Serialization of an Edge", "[Edge][load_minimal][save_minimal]") {
-  testRoundtrip(Edge(1337));
 }
-
-}  // namespace flatsurf::test
